backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
aiomysql
python-dotenv
pydantic
pydantic-core
starlette
aiosmtplib
email-validator
